feat: migrated components from decisiontree.py to a sub-package. also migrated the tests accordingly
