+++ src/random_tree_models/gradientboostedtrees.py
@@ -14,7 +14,7 @@ from sklearn.utils.validation import (
 )
 
 import random_tree_models.decisiontree as dtree
-from random_tree_models.scoring import MetricNames
+from random_tree_models.params import MetricNames
 
 
 class GradientBoostedTreesTemplate(base.BaseEstimator):
