+++ src/random_tree_models/randomforest.py
@@ -10,7 +10,7 @@ from sklearn.utils.validation import (
 )
 
 import random_tree_models.decisiontree as dtree
-from random_tree_models.scoring import MetricNames
+from random_tree_models.params import MetricNames
 
 
 class RandomForestTemplate(base.BaseEstimator):
