+++ tests/test_utils.py
@@ -3,26 +3,35 @@ import logging
 import pytest
 from pydantic import ValidationError
 
+import random_tree_models.params
 import random_tree_models.utils as utils
 
 
 def test_ColumnSelectionMethod():
     expected = ["ascending", "largest_delta", "random"]
-    assert list(utils.ColumnSelectionMethod.__members__.keys()) == expected
+    assert (
+        list(random_tree_models.params.ColumnSelectionMethod.__members__.keys())
+        == expected
+    )
 
 
 def test_ThresholdSelectionMethod():
     expected = ["bruteforce", "quantile", "random", "uniform"]
-    assert list(utils.ThresholdSelectionMethod.__members__.keys()) == expected
+    assert (
+        list(random_tree_models.params.ThresholdSelectionMethod.__members__.keys())
+        == expected
+    )
 
 
 # method, quantile, random_state, n_thresholds
 class TestThresholdSelectionParameters:
     def test_expected_okay(self):
-        params = utils.ThresholdSelectionParameters(
+        params = random_tree_models.params.ThresholdSelectionParameters(
             method="quantile", quantile=0.1, random_state=0, n_thresholds=100
         )
-        assert params.method == utils.ThresholdSelectionMethod.quantile
+        assert (
+            params.method == random_tree_models.params.ThresholdSelectionMethod.quantile
+        )
         assert params.quantile == 0.1
         assert params.random_state == 0
         assert params.n_thresholds == 100
@@ -30,7 +39,7 @@ class TestThresholdSelectionParameters:
 
     def test_method_fail(self):
         try:
-            _ = utils.ThresholdSelectionParameters(
+            _ = random_tree_models.params.ThresholdSelectionParameters(
                 method="wuppy", quantile=0.1, random_state=0, n_thresholds=100
             )
         except ValueError as ex:
@@ -44,7 +53,7 @@ class TestThresholdSelectionParameters:
     )
     def test_quantile(self, q: float, fail: bool):
         try:
-            _ = utils.ThresholdSelectionParameters(
+            _ = random_tree_models.params.ThresholdSelectionParameters(
                 method="quantile", quantile=q, random_state=0, n_thresholds=100
             )
         except ValueError as ex:
@@ -66,7 +75,7 @@ class TestThresholdSelectionParameters:
     )
     def test_random_state(self, random_state: int, fail: bool):
         try:
-            _ = utils.ThresholdSelectionParameters(
+            _ = random_tree_models.params.ThresholdSelectionParameters(
                 method="quantile",
                 quantile=0.1,
                 random_state=random_state,
@@ -95,7 +104,7 @@ class TestThresholdSelectionParameters:
     )
     def test_n_thresholds(self, n_thresholds: int, fail: bool):
         try:
-            _ = utils.ThresholdSelectionParameters(
+            _ = random_tree_models.params.ThresholdSelectionParameters(
                 method="quantile",
                 quantile=0.1,
                 random_state=42,
@@ -112,27 +121,31 @@ class TestThresholdSelectionParameters:
 
 
 def test_ColumnSelectionParameters():
-    params = utils.ColumnSelectionParameters(method="random", n_trials=10)
-    assert params.method == utils.ColumnSelectionMethod.random
+    params = random_tree_models.params.ColumnSelectionParameters(
+        method="random", n_trials=10
+    )
+    assert params.method == random_tree_models.params.ColumnSelectionMethod.random
     assert params.n_trials == 10
 
 
 class TestTreeGrowthParameters:
     def test_expected_okay(self):
-        params = utils.TreeGrowthParameters(
+        params = random_tree_models.params.TreeGrowthParameters(
             max_depth=10,
             min_improvement=0.0,
             lam=0.0,
             frac_subsamples=1.0,
             frac_features=1.0,
             random_state=0,
-            threshold_params=utils.ThresholdSelectionParameters(
+            threshold_params=random_tree_models.params.ThresholdSelectionParameters(
                 method="quantile",
                 quantile=0.1,
                 random_state=0,
                 n_thresholds=100,
             ),
-            column_params=utils.ColumnSelectionParameters(method="random", n_trials=10),
+            column_params=random_tree_models.params.ColumnSelectionParameters(
+                method="random", n_trials=10
+            ),
         )
         assert params.max_depth == 10
         assert params.min_improvement == 0.0
@@ -140,8 +153,13 @@ class TestTreeGrowthParameters:
         assert params.frac_subsamples == 1.0
         assert params.frac_features == 1.0
         assert params.random_state == 0
-        assert isinstance(params.threshold_params, utils.ThresholdSelectionParameters)
-        assert isinstance(params.column_params, utils.ColumnSelectionParameters)
+        assert isinstance(
+            params.threshold_params,
+            random_tree_models.params.ThresholdSelectionParameters,
+        )
+        assert isinstance(
+            params.column_params, random_tree_models.params.ColumnSelectionParameters
+        )
 
     @pytest.mark.parametrize(
         "frac_subsamples,fail",
@@ -155,7 +173,7 @@ class TestTreeGrowthParameters:
     )
     def test_frac_subsamples(self, frac_subsamples: float, fail: bool):
         try:
-            _ = utils.TreeGrowthParameters(
+            _ = random_tree_models.params.TreeGrowthParameters(
                 max_depth=10,
                 frac_subsamples=frac_subsamples,
             )
@@ -180,7 +198,7 @@ class TestTreeGrowthParameters:
     )
     def test_frac_features(self, frac_features: float, fail: bool):
         try:
-            _ = utils.TreeGrowthParameters(
+            _ = random_tree_models.params.TreeGrowthParameters(
                 max_depth=10,
                 frac_features=frac_features,
             )
@@ -195,7 +213,7 @@ class TestTreeGrowthParameters:
 
     def test_fail_if_max_depth_missing(self):
         with pytest.raises(ValidationError):
-            _ = utils.TreeGrowthParameters()  # type: ignore
+            _ = random_tree_models.params.TreeGrowthParameters()  # type: ignore
 
 
 def test_get_logger():
